Revamp state representation and network
The old state representation was very hard to effectively learn and
overly complex, attempting to shove the entire play history into a
set of integers that were difficult for the network to learn. Instead,
update the play state to be a set of one-hot (boolean) vectors for
each player in relative position to the picker indicating the cards
played on the current trick.

For past trick history, update the PPOAgent to use a feed-forward
network of 3 layers, followed by a LSTM to remember previous cards
played.

To ensure that each agent has access to the full trick history, add
an observe() method that propagates state updates to the LSTM without
expecting an action.

Add a few other minor state vector tweaks such as an explicit leaster
flag.