Slightly bias against leasters in reward
Leasters aren't tremendously fun, so slightly reduce
the game engine reward for them under all scenarios
to reduce their likelihood by the equivalent of 1 point.

Reduces liklihood of getting stuck in a local minimum of
always passing to get a leaster (where the chance of loss
is only -1 point per game, with a possible +4 point upside)

Under new reward structure, a standard picker loss (-2) should
be equivalent to a leaster loss, making picking and losing not
as unappealing if the alternative is a leaster.