Only penalize defenders playing trump when not forced
After 1 million training episodes of good progress, the agent
veered into always picking. With such a large negative reward
for leading trump, the reward structure discouraged defenders
with significantly stronger hands from punishing this behavior
and overall propagated a negative reward back to PASS.

Leading trump as a defender is almost always wrong in standard
play when the picker is expected to have more trump, but when
exploring this area of the search space in training, the large
negative reward likely contributed to algorithm divergence.