Initialize card embeddings with known suit, rank, and point values
Don't force the model to learn embeddings for known static values,
which dramatically slows learning. Instead initialize the embeddings
to normalized values in the initial embedding vectors. Learning can
still adjust these values and/or add additional features to extra
dimensions.

Increase card embedding size to 12 dimensions to allow the model the
possibility of encoding additional semantics in a few extra dimensions.

Reduce the learning rate of the card embeddings to 10% of the actor LR
to promote stability.