Switch OpenSkill rating to team based in pfsp.py
Instead of a different team per agent, use two teams for
non-leaster games (picking team and defending team). This
better mirrors the game dynamics and the OpenSkill rate API
and should produce less noisy skill ratings.

Since the teams are uneaven (2 vs 3 players) this will put
downward pressure on the rating sigma values, causing them
to hover around 0 over time. Their values are only useful in
a relative sense, so this should not be a problem for population
selection.