Add player point counts to critic and analyze output
Add a new head to the critic that predicts the points-per-player. This
value is entirely deterministic based on public play history
(only include the blind for the picking player) and should be learnable
with 100% accuracy. Adding this head should push the relevance of this
task down into the shared encoder and enable stronger strategic play in
the actor over time, equipping the agent to "know" how close to victory
each player/team is, similar to an expert human player counting points.

Expose this prediction in the /analyze app. Refactor React components
slightly to clean this up and move the secret partner prediction into
the expanded view.