Tweak reward shaping for slightly less bias
Now that the model is successfully distinguishing partner vs
defender behavior, drop the partner trump shaping a little to avoid
undue bias.

Drop the pick shaping slightly as well. With the current heuristic,
a hand of 2 queens and 1 jack has a stength value of 8 but is still
a somewhat risky pick, so best to have this shaping a bit lower.
Drop the negative reward shaping for poor hands to help balance.