Use state-conditioned logit for "all seen" highest trump class
Instead of a single bias parameter for the "all seen" highest
trump class corresponding to the dot product values over trump
cards, use linear projection of the "query" vector so that it
depends on the existing state. This should make the "all seen"
state much easier to learn and more accurate earlier in training
for a negligible increase in parameters.