Nudge leaster reward bonus slightly
Pick rate is hovering close to 50% and leaster rate below 0.5%. We are
slightly too biased against leasters currently. We would prefer the AI
to be biased against them, but it's a little too far in that
direction.

Increase the leaster reward bump 0.05 -> 0.1 so that a player
participating in a leaster and taking no tricks earns a final reward
just below 1 (-0.0033). The average non-winning player reward is
further reduced by the negative point rewards for any points they took.